Name spec by relationship type
Instead of naming specs by the ActiveRecord model name,
we name them by what type of underlying table to materialized
view mapping they are.

For instance, the basic spec we have now is a one to one mapping
because each row in the underlying table has exactly one row in the
materialized view.